credential widgets update
