+++ lib/home/tab_bar/credentials/widgets/credential_widget/credential_widget.dart
@@ -1,0 +1,17 @@
+export 'associated_wallet_widget.dart';
+export 'certificate_of_employment_widget.dart';
+export 'default_credential_subject_widget.dart';
+export 'ecole_42_learning_achievement_widget.dart';
+export 'email_pass_widget.dart';
+export 'identity_pass_widget.dart';
+export 'learning_achievement_widget.dart';
+export 'loyalty_card_widget.dart';
+export 'over18_widget.dart';
+export 'phone_pass_widget.dart';
+export 'professional_experience_assessment_widget.dart';
+export 'professional_skill_assessment_widget.dart';
+export 'professional_student_card_widget.dart';
+export 'resident_card_widget.dart';
+export 'self_issued_widget.dart';
+export 'student_card_widget.dart';
+export 'voucher_widget.dart';
