+++ lib/home/tab_bar/credentials/widgets/credential_widget/loyalty_card_widget.dart
