--- lib/home/tab_bar/credentials/widgets/default_credential_subject_widget.dart
@@ -1,388 +1,0 @@
-import 'package:altme/app/app.dart';
-import 'package:altme/home/home.dart';
-import 'package:altme/l10n/l10n.dart';
-import 'package:altme/theme/theme.dart';
-import 'package:credential_manifest/credential_manifest.dart';
-import 'package:flutter/material.dart';
-
-class DefaultCredentialSubjectDisplayInList extends StatelessWidget {
-  const DefaultCredentialSubjectDisplayInList({
-    Key? key,
-    required this.credentialModel,
-    this.descriptionMaxLine = 2,
-  }) : super(key: key);
-
-  final CredentialModel credentialModel;
-  final int descriptionMaxLine;
-
-  @override
-  Widget build(BuildContext context) {
-    final credential = Credential.fromJsonOrDummy(credentialModel.data);
-    final outputDescriptor =
-        credentialModel.credentialManifest?.outputDescriptors?.first;
-    // If outputDescriptor exist, the credential has a credential manifest
-    // telling us what to display
-
-    final backgroundColor = outputDescriptor == null
-        ? credentialModel
-            .credentialPreview.credentialSubjectModel.credentialSubjectType
-            .backgroundColor(credentialModel)
-        : getColorFromCredential(
-            outputDescriptor.styles?.background,
-            Colors.white,
-          );
-
-    late Widget descriptionWidget;
-
-    if (outputDescriptor == null) {
-      descriptionWidget = Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Expanded(
-            flex: 2,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                DisplayNameCard(
-                  credentialModel: credentialModel,
-                  style: Theme.of(context).textTheme.credentialTitle,
-                ),
-                const SizedBox(height: 5),
-                DisplayDescriptionCard(
-                  credentialModel: credentialModel,
-                  style: Theme.of(context).textTheme.credentialDescription,
-                  maxLines: descriptionMaxLine,
-                )
-              ],
-            ),
-          ),
-          Expanded(
-            flex: 1,
-            child: FractionallySizedBox(
-              heightFactor: 0.4,
-              child: DisplayIssuer(
-                issuer: credentialModel
-                    .credentialPreview.credentialSubjectModel.issuedBy!,
-              ),
-            ),
-          )
-        ],
-      );
-    }
-
-    if (outputDescriptor != null) {
-      final textColor =
-          getColorFromCredential(outputDescriptor.styles?.text, Colors.black);
-      descriptionWidget = Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Expanded(
-            flex: 2,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                if (outputDescriptor.display?.title != null)
-                  DisplayTitleWidget(
-                    displayMapping: outputDescriptor.display?.title,
-                    item: credentialModel,
-                    textColor: textColor,
-                  )
-                else
-                  const Text(''),
-                const SizedBox(height: 5),
-                DisplayDescriptionWidget(
-                  outputDescriptor.display?.description,
-                  credentialModel,
-                  textColor,
-                )
-              ],
-            ),
-          ),
-          Expanded(
-            flex: 1,
-            child: DisplayIssuanceDateWidget(
-              credentialModel.credentialPreview.issuanceDate,
-              textColor,
-            ),
-          )
-        ],
-      );
-    }
-
-    return CredentialContainer(
-      child: AspectRatio(
-        aspectRatio: 584 / 317,
-        child: Container(
-          decoration: BaseBoxDecoration(
-            color: backgroundColor,
-            shapeColor: Theme.of(context).colorScheme.documentShape,
-            value: 1,
-            anchors: const <Alignment>[Alignment.bottomRight],
-          ),
-          child: Row(
-            children: [
-              Expanded(
-                flex: 1,
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Expanded(
-                      child: HeroFix(
-                        tag: 'credential/${credentialModel.id}/icon',
-                        child: Container(
-                          alignment: Alignment.center,
-                          child: FractionallySizedBox(
-                            heightFactor: 0.4,
-                            child: FittedBox(
-                              child: Center(
-                                child: CredentialIcon(credential: credential),
-                              ),
-                            ),
-                          ),
-                        ),
-                      ),
-                    ),
-                    const SizedBox(height: 10),
-                    Expanded(
-                      child: Container(
-                        alignment: Alignment.center,
-                        child: FractionallySizedBox(
-                          heightFactor: 0.4,
-                          child: FittedBox(
-                            child: DisplayStatus(
-                              credentialModel: credentialModel,
-                              displayLabel: false,
-                            ),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              Expanded(
-                flex: 4,
-                child: Padding(
-                  padding: const EdgeInsets.only(top: 10, right: 10, bottom: 2),
-                  child: descriptionWidget,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class DefaultCredentialSubjectDisplayInSelectionList extends StatelessWidget {
-  const DefaultCredentialSubjectDisplayInSelectionList({
-    Key? key,
-    required this.credentialModel,
-    this.descriptionMaxLine = 2,
-  }) : super(key: key);
-
-  final CredentialModel credentialModel;
-  final int descriptionMaxLine;
-
-  @override
-  Widget build(BuildContext context) {
-    final outputDescriptor =
-        credentialModel.credentialManifest?.outputDescriptors?.first;
-    // If outputDescriptor exist, the credential has a credential manifest
-    // telling us what to display
-
-    final backgroundColor = outputDescriptor == null
-        ? credentialModel
-            .credentialPreview.credentialSubjectModel.credentialSubjectType
-            .backgroundColor(credentialModel)
-        : getColorFromCredential(
-            outputDescriptor.styles?.background,
-            Colors.white,
-          );
-
-    late Widget descriptionWidget;
-
-    if (outputDescriptor == null) {
-      descriptionWidget = Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Expanded(
-            flex: 2,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                DisplayNameCard(
-                  credentialModel: credentialModel,
-                  style: Theme.of(context).textTheme.credentialTitle,
-                ),
-                const SizedBox(height: 5),
-                DisplayDescriptionCard(
-                  credentialModel: credentialModel,
-                  style: Theme.of(context).textTheme.credentialDescription,
-                  maxLines: descriptionMaxLine,
-                )
-              ],
-            ),
-          ),
-          Expanded(
-            flex: 1,
-            child: FractionallySizedBox(
-              heightFactor: 0.4,
-              child: DisplayIssuer(
-                issuer: credentialModel
-                    .credentialPreview.credentialSubjectModel.issuedBy!,
-              ),
-            ),
-          )
-        ],
-      );
-    }
-
-    if (outputDescriptor != null) {
-      final textColor =
-          getColorFromCredential(outputDescriptor.styles?.text, Colors.black);
-      descriptionWidget = Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Expanded(
-            flex: 2,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                if (outputDescriptor.display?.title != null)
-                  DisplayTitleWidget(
-                    displayMapping: outputDescriptor.display?.title,
-                    item: credentialModel,
-                    textColor: textColor,
-                  )
-                else
-                  const Text(''),
-                DisplayDescriptionWidget(
-                  outputDescriptor.display?.description,
-                  credentialModel,
-                  textColor,
-                )
-              ],
-            ),
-          ),
-          Expanded(
-            flex: 1,
-            child: DisplayIssuanceDateWidget(
-              credentialModel.credentialPreview.issuanceDate,
-              textColor,
-            ),
-          )
-        ],
-      );
-    }
-
-    return CredentialContainer(
-      child: AspectRatio(
-        aspectRatio: 584 / 317,
-        child: Container(
-          // margin: const EdgeInsets.symmetric(vertical: 4.0),
-          decoration: BaseBoxDecoration(
-            borderRadius: BorderRadius.circular(20),
-            color: backgroundColor,
-            shapeColor: Theme.of(context).colorScheme.documentShape,
-            value: 1,
-            anchors: const <Alignment>[
-              Alignment.bottomRight,
-            ],
-          ),
-          child: Padding(
-            padding: const EdgeInsets.all(20),
-            child: descriptionWidget,
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class DefaultCredentialSubjectDisplayDetail extends StatelessWidget {
-  const DefaultCredentialSubjectDisplayDetail({
-    Key? key,
-    required this.credentialModel,
-  }) : super(key: key);
-
-  final CredentialModel credentialModel;
-
-  @override
-  Widget build(BuildContext context) {
-    final outputDescriptor =
-        credentialModel.credentialManifest?.outputDescriptors;
-    // If outputDescriptor exist, the credential has a credential manifest
-    // telling us what to display
-    if (outputDescriptor == null) {
-      final l10n = context.l10n;
-
-      return AspectRatio(
-        aspectRatio: 584 / 317,
-        child: CredentialBackground(
-          credentialModel: credentialModel,
-          child: Column(
-            children: [
-              Padding(
-                padding: const EdgeInsets.all(8),
-                child: DisplayNameCard(
-                  credentialModel: credentialModel,
-                  style: Theme.of(context).textTheme.credentialTitle,
-                ),
-              ),
-              Padding(
-                padding: const EdgeInsets.all(8),
-                child: DisplayDescriptionCard(
-                  credentialModel: credentialModel,
-                  style: Theme.of(context).textTheme.credentialDescription,
-                ),
-              ),
-              if (credentialModel.credentialPreview.evidence.first.id != '')
-                Padding(
-                  padding: const EdgeInsets.all(8),
-                  child: Row(
-                    children: [
-                      Text(
-                        '${l10n.evidenceLabel}: ',
-                        style: Theme.of(context).textTheme.credentialFieldTitle,
-                      ),
-                      Flexible(
-                        child: InkWell(
-                          onTap: () => LaunchUrl.launch(
-                            credentialModel.credentialPreview.evidence.first.id,
-                          ),
-                          child: Padding(
-                            padding: const EdgeInsets.all(8),
-                            child: Row(
-                              children: [
-                                Text(
-                                  credentialModel
-                                      .credentialPreview.evidence.first.id,
-                                  style: Theme.of(context)
-                                      .textTheme
-                                      .credentialFieldDescription,
-                                  maxLines: 5,
-                                  overflow: TextOverflow.fade,
-                                  softWrap: true,
-                                ),
-                              ],
-                            ),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
-                )
-              else
-                const SizedBox.shrink()
-            ],
-          ),
-        ),
-      );
-    } else {
-      return OutputDescriptorWidget(outputDescriptor, credentialModel);
-    }
-  }
-}
