+++ lib/home/tab_bar/credentials/widgets/display_description_card.dart
@@ -21,7 +21,6 @@ class DisplayDescriptionCard extends StatelessWidget {
     return MyText(
       nameValue,
       maxLines: maxLines,
-      overflow: TextOverflow.fade,
       style: style,
     );
   }
