+++ lib/home/tab_bar/credentials/widgets/widgets.dart
@@ -1,11 +1,13 @@
-export 'associated_wallet_widget.dart';
 export 'card_animation.dart';
-export 'certificate_of_employment_widget.dart';
 export 'credential_background.dart';
 export 'credential_container.dart';
 export 'credential_image.dart';
+export 'credential_manifest_display_descriptor.dart';
+export 'credential_manifest_selection_display_descriptor.dart';
 export 'credential_selection_padding.dart';
-export 'default_credential_subject_widget.dart';
+export 'credential_widget/credential_widget.dart';
+export 'default_display_descriptor.dart';
+export 'default_selection_display_descriptor.dart';
 export 'display_description_card.dart';
 export 'display_description_widget.dart';
 export 'display_detail.dart';
@@ -19,23 +21,8 @@ export 'display_properties_widget.dart';
 export 'display_signature.dart';
 export 'display_status.dart';
 export 'display_title_widget.dart';
-export 'ecole_42_learning_achievement_widget.dart';
-export 'email_pass_widget.dart';
 export 'home_credential_item.dart';
-export 'identity_pass_widget.dart';
 export 'labeled_display_mapping_widget.dart';
-export 'learning_achievement_widget.dart';
 export 'list_item.dart';
-export 'loyalty_card_widget.dart';
-export 'output_descriptor_widget.dart';
-export 'over18_widget.dart';
-export 'phone_pass_widget.dart';
-export 'professional_experience_assessment_widget.dart';
-export 'professional_skill_assessment_widget.dart';
-export 'professional_student_card_widget.dart';
-export 'resident_card_widget.dart';
-export 'self_issued_widget.dart';
 export 'skills_list_display.dart';
 export 'star_rating.dart';
-export 'student_card_widget.dart';
-export 'voucher_widget.dart';
